+++ packages/front-end/app/login/_type/OAuth2DataType.ts
@@ -1,0 +1,13 @@
+export interface OAuth2DataType {
+    pathname: string;
+    query: {
+        client_id: string;
+        redirect_uri: string;
+        scope?: string;
+        state?: string;
+    }
+    display: {
+        text: string;
+        iconURL: string;
+    }
+}
