+++ packages/front-end/app/login/_type/ProviderEnum.ts
@@ -1,0 +1,1 @@
+export type ProviderEnum = "GOOGLE" | "NAVER" | "KAKAO";
